Road Data for Curitiba

A dataset containing road information for Curitiba, Brazil.

Usage

ippuc_road_axis

Format

A data frame and sf object with 40,355 observations and 22 variables:

CHAVE

chr. The unique key identifier for each road segment.

CODVIA

chr. The code of the road.

NOINICIO

chr. The starting number of the road segment.

NOFIM

chr. The ending number of the road segment.

STATUS

chr. The status of the road segment (e.g., "OFICIAL").

NMVIA

chr. The name of the road.

NMVIA_ANT

chr. The former name of the road, if any.

SVIARIO

chr. The classification of the road.

SVIARIOLEG

chr. The legal reference for the road classification.

HIERARQUIA

chr. The hierarchy level of the road.

STATUS_DEM

chr. The official status of the road.

COD_BAIRRO

num. The code of the neighborhood

BAIRRO_E

chr. The name of the neighborhood on the left side of the road.

BAIRRO_D

chr. The name of the neighborhood on the right side of the road.

COD_REG

num. The code of the region.

REGIONAL_E

chr. The name of the regional administration on the left side of the road.

REGIONAL_D

chr. The name of the regional administration on the right side of the road.

CEP_E

chr. The postal code on the left side of the road.

CEP_D

chr. The postal code on the right side of the road.

CEP_INFO

chr. Information about the postal code.

CEP_FONTE

chr. The source of the postal code information.

geometry

sfc_MULTILINESTRING. The spatial geometry of the road segment.

Source

Provided by the municipal authorities of Curitiba (IPPUC).