Example NDS-BR Spatial Dataset

A dataset containing trip information of the NDS-BR research project

Usage

ndsbr_data_sf

Format

A data frame and sf object with 18,012 observations and 33 variables:

DRIVER

chr. The identifier for the driver.

DAY

chr. The date of the trip.

DAY_CORRIGIDO

chr. The corrected date of the trip.

03:00:00

chr. The timestamp

TRIP

dbl. The trip number.

ID

chr. The trip identifier.

PR

chr. The timestamp in GMT-3 timezone

H

dbl. The hour component of the timestamp.

M

dbl. The minute component of the timestamp.

S

dbl. The second component of the timestamp.

TIME_ACUM

dbl. The accumulated time of the trip in seconds.

SPD_MPH

dbl. The speed in miles per hour.

SPD_KMH

dbl. The speed in kilometers per hour.

ACEL_MS2

dbl. The acceleration in meters per second squared.

HEADING

dbl. The heading or direction of travel.

ALTITUDE_FT

dbl. The altitude in feet.

VALID_TIME

chr. Indicates whether the part of the trip is valid ("Yes" or "No").

TIMESTAMP_GPS

chr. Indicates whether the GPS timestamp is valid ("No").

CPOOL

chr. Indicates whether the trip is carpooling ("Not carpooling").

CPOOLING_CHECKED

chr. Indicates whether carpooling was checked ("Not carpooling").

WSB

chr. Indicates whether the driver was using seatbelt ("Yes").

UMP_YN

chr. Indicates whether the driver was using cellphone ("No").

UMP

chr. Indicates whether the driver was using cellphone ("0").

PICK_UP

chr. Indicates the pick-up cellphone quantity ("0").

ACTION

chr. Indicates the action taken during the usage of cellphone (NA).

GPS_FILE

chr. The GPS file identifier (NA).

CIDADE

chr. The city name ("Curitiba").

BAIRRO

chr. The neighborhood name ("JARDIM DAS AMÉRICAS").

NOME_RUA

chr. The street name ("NPI").

HIERARQUIA_CWB

chr. The road hierarchy according to the city of Curitiba ("NPI").

HIERARQUIA_CTB

chr. The road hierarchy according to the Brazilian Traffic Code ("NPI").

LIMITE_VEL

chr. The speed limit ("NPI").

geometry

POINT. The spatial geometry of the trip location.

Source

NDS-BR researchers.